Add inference for definition arguments
