+++ compiler/hash-typecheck/src/new/ops/mod.rs
@@ -1,15 +1,14 @@
 //! Contains various TC-related operations, which form the core of the
 //! typechecker.
 use self::{
-    bootstrap::BootstrapOps, check::CheckOps, context::ContextOps, infer::InferOps,
-    normalise::NormalisationOps, substitutions::SubstituteOps, unify::UnifyOps,
+    bootstrap::BootstrapOps, check::CheckOps, context::ContextOps, normalise::NormalisationOps,
+    substitutions::SubstituteOps, unify::UnifyOps,
 };
 
 pub mod bootstrap;
 pub mod check;
 pub mod common;
 pub mod context;
-pub mod infer;
 pub mod normalise;
 pub mod substitutions;
 pub mod unify;
@@ -29,9 +28,8 @@ macro_rules! ops {
 }
 
 ops! {
-  infer_ops: InferOps,
-  unify_ops: UnifyOps,
   check_ops: CheckOps,
+  unify_ops: UnifyOps,
   context_ops: ContextOps,
   bootstrap_ops: BootstrapOps,
   normalisation_ops: NormalisationOps,
